+++ src/component/CardAddButton/styled.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 
 const St = {
-  AddButtonSection: styled.section`
+  AddButtonSection: styled.li`
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -20,6 +20,7 @@ const St = {
 
     background: #e5e5e5;
     border-radius: 5px;
+    border: none;
 
     font: 700 30px/35px "Roboto";
     color: #575757;
