+++ src/component/CardListPage/CardListHeader/styled.tsx
@@ -1,11 +1,6 @@
 import styled from "styled-components";
 
 const St = {
-  Header: styled.header`
-    align-items: center;
-    display: flex;
-    gap: 18px;
-  `,
   Title: styled.h1`
     align-items: center;
     font: 400 17px "Roboto";
